Added experience and mountain location.

+++ public/shop.js
@@ -58,11 +58,13 @@ function buy1(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal1").innerHTML;
-		vals.inventory.push({"tag":"Leather Armor","myType":"armor","meleeBonus":0,"rangedBonus":0,"magicBonus":0,"defenseBonus":2});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Leather Armor");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem1").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal1").innerHTML;
+			vals.inventory.push({"tag":"Leather Armor","myType":"armor","meleeBonus":0,"rangedBonus":0,"magicBonus":0,"defenseBonus":2});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Leather Armor");
+		}
 	}
 }
 
@@ -72,11 +74,13 @@ function buy2(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal2").innerHTML;
-		vals.inventory.push({"tag":"Wooden Dagger","myType":"weapon","meleeBonus":2,"rangedBonus":0,"magicBonus":0,"defenseBonus":0});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Wooden Dagger");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem2").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal2").innerHTML;
+			vals.inventory.push({"tag":"Wooden Dagger","myType":"weapon","meleeBonus":2,"rangedBonus":0,"magicBonus":0,"defenseBonus":0});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Wooden Dagger");
+		}
 	}
 
 }
@@ -87,11 +91,13 @@ function buy3(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal3").innerHTML;
-		vals.inventory.push({"tag":"Knight Armor","myType":"armor", "meleeBonus":0,"rangedBonus":0,"magicBonus":0,"defenseBonus":3});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Knight Armor");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem3").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal3").innerHTML;
+			vals.inventory.push({"tag":"Knight Armor","myType":"armor", "meleeBonus":0,"rangedBonus":0,"magicBonus":0,"defenseBonus":3});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Knight Armor");
+		}
 	}
 }
 
@@ -101,11 +107,13 @@ function buy4(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal4").innerHTML;
-		vals.inventory.push({"tag":"Knight Sword","myType":"weapon", "meleeBonus":3,"rangedBonus":0,"magicBonus":0,"defenseBonus":0});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Knight Sword");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem4").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal4").innerHTML;
+			vals.inventory.push({"tag":"Knight Sword","myType":"weapon", "meleeBonus":3,"rangedBonus":0,"magicBonus":0,"defenseBonus":0});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Knight Sword");
+		}
 	}
 }
 
@@ -115,11 +123,13 @@ function buy5(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal5").innerHTML;
-		vals.inventory.push({"tag":"Wizard Robe","myType":"armor", "meleeBonus":0,"rangedBonus":0,"magicBonus":2,"defenseBonus":2});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Wizard Robe");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem5").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal5").innerHTML;
+			vals.inventory.push({"tag":"Wizard Robe","myType":"armor", "meleeBonus":0,"rangedBonus":0,"magicBonus":2,"defenseBonus":2});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Wizard Robe");
+		}
 	}
 }
 
@@ -129,11 +139,13 @@ function buy6(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal6").innerHTML;
-		vals.inventory.push({"tag":"Wizard Staff","myType":"weapon", "meleeBonus":0,"rangedBonus":0,"magicBonus":4,"defenseBonus":0});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Wizard Staff");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem6").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal6").innerHTML;
+			vals.inventory.push({"tag":"Wizard Staff","myType":"weapon", "meleeBonus":0,"rangedBonus":0,"magicBonus":4,"defenseBonus":0});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Wizard Staff");
+		}
 	}
 }
 
@@ -143,11 +155,13 @@ function buy7(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal7").innerHTML;
-		vals.inventory.push({"tag":"Ninja Garment","myType":"armor", "meleeBonus":3,"rangedBonus":3,"magicBonus":3,"defenseBonus":2});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Ninja Garment");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem7").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal7").innerHTML;
+			vals.inventory.push({"tag":"Ninja Garment","myType":"armor", "meleeBonus":3,"rangedBonus":3,"magicBonus":3,"defenseBonus":2});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Ninja Garment");
+		}
 	}
 }
 
@@ -157,141 +171,194 @@ function buy8(){
 		return;
 	}
 	else {
-		vals.gold = currentGold - document.getElementById("shopItemVal8").innerHTML;
-		vals.inventory.push({"tag":"Throwing Stars","myType":"weapon", "meleeBonus":0,"rangedBonus":5,"magicBonus":0,"defenseBonus":0});
-		localStorage.setItem('userData', JSON.stringify(vals));
-		window.location.href = "/shop.html";
-		alert("Purchased Throwing Stars");
+		if (confirm("Confirm purchase of '" + document.getElementById("shopItem8").innerHTML + "'?")){
+			vals.gold = currentGold - document.getElementById("shopItemVal8").innerHTML;
+			vals.inventory.push({"tag":"Throwing Stars","myType":"weapon", "meleeBonus":0,"rangedBonus":5,"magicBonus":0,"defenseBonus":0});
+			localStorage.setItem('userData', JSON.stringify(vals));
+			window.location.href = "/shop.html";
+			alert("Purchased Throwing Stars");
+		}
 	}
 }
 
 function slotSell1(){
 	var temp = vals.inventory[0];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(0, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[0].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(0, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell2(){
 	var temp = vals.inventory[1];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(1, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[1].tag + "'?")){
+
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(1, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}	
 
 }
 
 function slotSell3(){
 	var temp = vals.inventory[2];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(2, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[2].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(2, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell4(){
 	var temp = vals.inventory[3];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(3, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[3].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(3, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell5(){
 	var temp = vals.inventory[4];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(4, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[4].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(4, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell6(){
 	var temp = vals.inventory[5];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(5, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[5].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(5, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell7(){
 	var temp = vals.inventory[6];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(6, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[6].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(6, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell8(){
 	var temp = vals.inventory[7];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(7, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[7].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(7, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell9(){
 	var temp = vals.inventory[8];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(8, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[8].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(8, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
 
 function slotSell10(){
 	var temp = vals.inventory[9];
-	for (i = 1; i <= shopCount; i++){
-		if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
-			vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
-	}
 
-	vals.inventory.splice(9, 1);
-	localStorage.setItem('userData', JSON.stringify(vals));
-	window.location.href = "/shop.html";
+	if (confirm("Are you sure you want to sell '" + vals.inventory[9].tag + "'?")){
+
+		for (i = 1; i <= shopCount; i++){
+			if (temp.tag == document.getElementById("shopItem" + i.toString()).innerHTML)
+				vals.gold = vals.gold + parseInt(document.getElementById("shopItemVal" + i.toString()).innerHTML);
+		}
+
+		vals.inventory.splice(9, 1);
+		localStorage.setItem('userData', JSON.stringify(vals));
+		window.location.href = "/shop.html";
+
+	}
 
 }
